pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linodecli-build"
version = "0.2.0"
description = "linode-cli plugin for deploying applications with declarative templates"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Linode", email = "support@linode.com" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML>=6.0",
    "textual>=0.47.0",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.urls]
Homepage = "https://github.com/linode/linode-cli-build"

[tool.setuptools.packages.find]
where = ["."]
include = ["linodecli_build*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"linodecli_build" = ["templates/**/*.yml", "tui/**/*.tcss"]
